code init

+++ core/server/models/index.js
@@ -11,26 +11,38 @@ const db = {};
 
 // 实例化sequelize
 config.database.operatorsAliases = Sequelize.Op; // 使用symbol操作符，防止注入
-const sequelize = new Sequelize(config.database);
+//const sequelize = new Sequelize(config.database);
+const sequelize = new Sequelize('test', 'root', '', config.database);
 
 
-fs.readdirSync(__dirname).filter(file=> {
-    // 排除index.js
-    return (file.indexOf('.') !== 0) && (file !== 'index.js');
-}).forEach(file=> {
-    const model = sequelize.import(path.join(__dirname, file));
+/**
+ * 动态获取models
+ */
+function getFileSync(rootPath) {
+    fs.readdirSync(rootPath).filter(file=> {
+        return file !== 'index.js';
+    }).forEach(file=> {
+        const _path = path.join(rootPath, file);
+        const info = fs.statSync(_path);
+        if(info.isDirectory()) { // 如果是目录继续编辑
+            getFileSync(_path);
+            return;
+        }
 
-    // 如果表未创建则先创建table
-    model.sync({force: false}).then(()=> {
-        logger.data('table:%s created success...', model.tableName);
-    }).catch(error=> {
-        logger.error('table:%s created failure...', model.tableName, error)
+        const model = sequelize.import(_path);
+        // 如果表未创建则先创建table
+        model.sync({force: false}).then(()=> {
+            logger.data('table:%s created success...', model.tableName);
+        }).catch(error=> {
+            logger.error('table:%s created failure...', model.tableName, error)
+        });
+
+        db[model.name] = model;
     });
+}
+getFileSync(__dirname);
 
-    db[model.name] = model;
-});
-
-db.sequelize = sequelize;
-db.Sequelize = Sequelize;
+db.sequelize = sequelize; // 实例
+db.Sequelize = Sequelize; // 类
 
 module.exports = db;
